remove log density file to avoid confusion and move posterior approximation utils to wrapper utils file
